web-app-deprecated/requirements.txt
streamlit>=1.41.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
